Expose inline contract via Contract.inline(); add contract-first sampling()
Replaces the standalone InlineContract type with a Contract.inline()
factory and a nested Contract.Inline builder. The naming encodes the
principle: an inline contract is exactly the operational layer
(Contract) with no identity, which is why it offers only contractual
criteria and no empirical path — graduating to a named ServiceContract
(the metadata/identity layer) is what unlocks empirical baselines.

The factor-less call site now reads contract-first via .sampling(n,
inputs), which returns Sampling<NoFactors,IT,OT>, instead of nesting
the contract as a parameter of Sampling.of(...). .build() remains for
the factored explore/optimize path.
